requirements.txt
torch>=2.2.0
diffusers>=0.27.0
transformers>=4.40.0
accelerate>=0.29.0
safetensors>=0.4.3
xformers>=0.0.25
sentencepiece>=0.2.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pillow>=10.3.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
python-dotenv>=1.0.0
